+++ quickbooks/objects/creditmemo.py
@@ -1,5 +1,3 @@
-from six import python_2_unicode_compatible
-
 from quickbooks.objects.detailline import SalesItemLine, SubtotalLine, DiscountLine, DescriptionOnlyLine, DetailLine
 from .base import Address, EmailAddress, Ref, CustomField, CustomerMemo, QuickbooksManagedObject, \
     LinkedTxnMixin, QuickbooksTransactionEntity
@@ -7,7 +5,6 @@ from .tax import TxnTaxDetail
 from ..mixins import DeleteMixin
 
 
-@python_2_unicode_compatible
 class CreditMemo(DeleteMixin, QuickbooksTransactionEntity, QuickbooksManagedObject, LinkedTxnMixin):
     """
     QBO definition: The CreditMemo is a financial transaction representing a refund or credit of payment or part
@@ -73,3 +70,11 @@ class CreditMemo(DeleteMixin, QuickbooksTransactionEntity, QuickbooksManagedObje
 
     def __str__(self):
         return str(self.TotalAmt)
+
+    def to_ref(self):
+        ref = Ref()
+
+        ref.type = self.qbo_object_name
+        ref.value = self.Id
+
+        return ref