+++ quickbooks/objects/employee.py
@@ -1,8 +1,6 @@
-from six import python_2_unicode_compatible
 from .base import Address, PhoneNumber, QuickbooksManagedObject, QuickbooksTransactionEntity, Ref
 
 
-@python_2_unicode_compatible
 class Employee(QuickbooksManagedObject, QuickbooksTransactionEntity):
     """
     QBO definition: Employee represents the people who are working for the company.
@@ -37,6 +35,7 @@ class Employee(QuickbooksManagedObject, QuickbooksTransactionEntity):
         self.BillableTime = False
 
         self.PrimaryAddr = None
+        self.PrimaryPhone = None
 
     def __str__(self):
         return self.DisplayName