+++ quickbooks/objects/taxagency.py
@@ -1,8 +1,6 @@
-from six import python_2_unicode_compatible
 from .base import QuickbooksTransactionEntity, QuickbooksManagedObject
 
 
-@python_2_unicode_compatible
 class TaxAgency(QuickbooksManagedObject, QuickbooksTransactionEntity):
     """
     QBO definition: Tax Agency is an entity that is associated with a tax rate and identifies the agency to which that tax rate